+++ stt/run_1.sh
@@ -1,7 +1,13 @@
 # Create folder ./cache if not exist
 mkdir -p cache
-sudo docker stop stt_1
-sudo docker rm stt_1
+
+# Check if the container exists, then stop and remove it
+if [ $(sudo docker ps -a -f name=stt_1 --format '{{.Names}}') = 'stt_1' ]; then
+    sudo docker stop stt_1
+    sudo docker rm stt_1
+fi
+
+# Run the Docker container
 sudo docker run \
     --gpus '"device=1"' \
     --restart always \
@@ -12,4 +18,5 @@ sudo docker run \
     -v $(pwd)/config_1.json:/app/config.json \
     -v $(pwd)/credentials_full.json:/app/credentials_full.json \
     -e GOOGLE_APPLICATION_CREDENTIALS=/app/credentials_full.json \
+    -e CUDA_LAUNCH_BLOCKING=1 \
     -t stt
